[Ref] IT Compression: Silence -Wconversion warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@18563 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/ITCompression.cpp
@@ -277,7 +277,7 @@ void ITCompression::WriteBits(int8 width, int v)
 {
 	while(width > remBits)
 	{
-		byteVal |= (v << bitPos);
+		byteVal |= static_cast<uint8>(v << bitPos);
 		width -= remBits;
 		v >>= remBits;
 		bitPos = 0;
@@ -288,7 +288,7 @@ void ITCompression::WriteBits(int8 width, int v)
 
 	if(width > 0)
 	{
-		byteVal |= (v & ((1 << width) - 1)) << bitPos;
+		byteVal |= static_cast<uint8>((v & ((1 << width) - 1)) << bitPos);
 		remBits -= width;
 		bitPos += width;
 	}
